+++ app/Permission.php
@@ -74,6 +74,54 @@ class Permission extends \Spatie\Permission\Models\Permission
             'add_permissions',
             'edit_permissions',
             'delete_permissions',
+
+            'view_logs',
+        ];
+    }
+
+    public static function modulePermission()
+    {
+        return [
+            'view_satker',
+            'add_satker',
+            'edit_satker',
+            'delete_satker',
+
+            'view_jabatan',
+            'add_jabatan',
+            'edit_jabatan',
+            'delete_jabatan',
+
+            'view_jenker',
+            'add_jenker',
+            'edit_jenker',
+            'delete_jenker',
+
+            'view_pekerjaan',
+            'add_pekerjaan',
+            'edit_pekerjaan',
+            'delete_pekerjaan',
+
+            'view_kontrak',
+            'add_kontrak',
+            'edit_kontrak',
+            'delete_kontrak',
+
+            'view_lahan',
+            'add_lahan',
+            'edit_lahan',
+            'delete_lahan',
+
+            'view_pengguna',
+            'add_pengguna',
+            'edit_pengguna',
+            'delete_pengguna',
+
+            'view_pelaksanaan',
+            'add_pelaksanaan',
+            'edit_pelaksanaan',
+            'delete_pelaksanaan',
+
         ];
     }
 }