+++ database/migrations/2020_10_21_002658_create_dokumentasi_sosialisasis_table.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateDokumentasiSosialisasisTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('dokumentasi_sosialisasis', function (Blueprint $table) {
+            $table->id();
+            $table->string('uuid')->unique();
+            $table->bigInteger('pekerjaan_id')->nullable();
+            $table->bigInteger('sosialisasi_id')->nullable();
+            $table->string('filename')->nullable();
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('dokumentasi_sosialisasis');
+    }
+}