+++ routes/web.php
@@ -34,10 +34,16 @@ Route::group(['prefix' => 'backoffice', 'middleware' => ['auth']], function() {
     Route::resource('satker', 'SatkerController');
     Route::resource('jabatan', 'JabatanController');
     Route::resource('pekerjaan', 'PekerjaanController');
-    Route::resource('jenker', 'JenisPekerjaanCOntroller');
+    Route::resource('jenker', 'JenisPekerjaanController');
+    Route::resource('pengguna', 'PenggunaController');
+    Route::resource('pelaksanaan', 'PelaksanaanController');
 
     // user Profile
     Route::get('profile', ['as'=>'profile','uses'=>'UserController@profile']);
     Route::patch('profile/{user}/update',['as' => 'profile.update', 'uses' => 'UserController@ProfileUpdate']);
     Route::patch('profile/{user}/password',['as' => 'profile.password','uses' => 'UserController@ChangePassword']);
+
+    // upload dokumentasi pelaksanaan
+    Route::get('pelaksanaan/{uuid}/dokumentasi',['as' => 'pelaksanaan.dokumentasi','uses'=> 'PelaksanaanController@getUpload']);
+    Route::post('pelaksanaan/dokumentasi/upload',['as' => 'pelaksanaan.upload','uses'=> 'PelaksanaanController@uploadImages']);
 });