+++ CHANGELOG.md
@@ -2,6 +2,68 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0-alpha.4](https://git.fiotech.co/wbaindonesia/project-reporting/compare/v1.1.0-alpha.3...v1.1.0-alpha.4) (2020-10-24)
+
+
+### Features
+
+* Adding new basic data table seeder ([49cfa3f](https://git.fiotech.co/wbaindonesia/project-reporting/commit/49cfa3f3cfca6b09e57cdfb21b9517eaf12fc0cc))
+* Adding pelaksaan keuangan edit form ([4b3d70a](https://git.fiotech.co/wbaindonesia/project-reporting/commit/4b3d70aa345c4c0515ad133644f370a44649d13d))
+* Change to pelaksanaan and adding  pelaksanaan keuangan view ([4b31d8e](https://git.fiotech.co/wbaindonesia/project-reporting/commit/4b31d8e3b662ec7d95c0557389a6879e3604de75))
+* Move fisiks table to pelaksanaans so it provide pelaksanaan keuangan ([f84a5f3](https://git.fiotech.co/wbaindonesia/project-reporting/commit/f84a5f34cbf57b92974d29ae624b50e1daa9bbd4))
+* Now support edit pelaksaaan keuangan ([787fa4a](https://git.fiotech.co/wbaindonesia/project-reporting/commit/787fa4a95f06b3de80eac7acab9fb77b8ce3a7f5))
+* Now support to adding pelaksanaan keuangan ([aadcbe2](https://git.fiotech.co/wbaindonesia/project-reporting/commit/aadcbe2b971f21d73e7a0c0eb52c0e1013b9f108))
+
+## [1.1.0-alpha.3](https://git.fiotech.co/wbaindonesia/project-reporting/compare/v1.1.0-alpha.2...v1.1.0-alpha.3) (2020-10-22)
+
+
+### Features
+
+* Add model pelaksanaan fisik ([9d9d642](https://git.fiotech.co/wbaindonesia/project-reporting/commit/9d9d642cd882ac0f50f9cf113262b1df4d0d34e4))
+* Add module permission array ([c6b7886](https://git.fiotech.co/wbaindonesia/project-reporting/commit/c6b7886ae33c17a4d80d2c390f95e9343bee24e3))
+* Add module permission seeder ([881639c](https://git.fiotech.co/wbaindonesia/project-reporting/commit/881639c4885e1769d702acf0cf3b6a67c1f26f15))
+* Add views for pelaksanaan fisik ([a8bac35](https://git.fiotech.co/wbaindonesia/project-reporting/commit/a8bac35505e6dcaf7798a034c91aebbb1bfb1d01))
+* Adding logic fot pelaksanaan fisik ([d6dea0f](https://git.fiotech.co/wbaindonesia/project-reporting/commit/d6dea0f7e22b31c6846c2f4fad5c83d0e9cf0513))
+* Adding menu pelaksanaan pekerjaan fisik ([8b80147](https://git.fiotech.co/wbaindonesia/project-reporting/commit/8b80147faa841afedd175845a99109533ef84a02))
+
+
+### Bug Fixes
+
+* Change tahun datatype to year ([a02c6a8](https://git.fiotech.co/wbaindonesia/project-reporting/commit/a02c6a8a9fd6e97a4a6eb356ad389922e3ad5084))
+* Now have validation before store ([d62c94f](https://git.fiotech.co/wbaindonesia/project-reporting/commit/d62c94f8d5ec00dc7d3cbcb6c954d48499351fc7))
+
+## [1.1.0-alpha.2](https://git.fiotech.co/wbaindonesia/project-reporting/compare/v1.1.0-alpha.1...v1.1.0-alpha.2) (2020-10-20)
+
+
+### Features
+
+* Adding jabatan table seeder ([9026193](https://git.fiotech.co/wbaindonesia/project-reporting/commit/902619315d1e68bd6a799ef4a44d7ec09f71abf4))
+* Adding jabatan_urutan column in users table ([0eef687](https://git.fiotech.co/wbaindonesia/project-reporting/commit/0eef6870bc506e56d2428f83514ae4b9dd1b6e4b))
+* Adding jenis pekerjaan table seeder ([fcf4a5c](https://git.fiotech.co/wbaindonesia/project-reporting/commit/fcf4a5c81bf4b89204d2936bb2c37d852b248c78))
+* Adding pengguna controller module ([2e42e30](https://git.fiotech.co/wbaindonesia/project-reporting/commit/2e42e308ab16235a8acf03dba5065e408a1f9416))
+* Adding pengguna route ([65db99a](https://git.fiotech.co/wbaindonesia/project-reporting/commit/65db99adefe482fb793733acfa188ea1c3ffacdd))
+* Adding relation to satker and jabatan ([7bc8c2b](https://git.fiotech.co/wbaindonesia/project-reporting/commit/7bc8c2b2248d4507fc98ee82c4b907a97566a2cb))
+* Adding satker table seeder ([c4a7a9f](https://git.fiotech.co/wbaindonesia/project-reporting/commit/c4a7a9fc4bcacc531387751992e839de1d34c920))
+* Adding views modul pengguna ([719046e](https://git.fiotech.co/wbaindonesia/project-reporting/commit/719046e045af251343d0979044e3c54f6b4c9ef1))
+
+
+### Bug Fixes
+
+* Change wilayah option to dropdown ([d45467c](https://git.fiotech.co/wbaindonesia/project-reporting/commit/d45467c1cf8b3b106e12bf6332258fc4584ff75f))
+* Menu layout in sidebar ([cefb5be](https://git.fiotech.co/wbaindonesia/project-reporting/commit/cefb5be058face743cbb60342a54d9060af35274))
+
+## [1.1.0-alpha.1](https://git.fiotech.co/wbaindonesia/project-reporting/compare/v1.1.0-alpha.0...v1.1.0-alpha.1) (2020-10-20)
+
+
+### Features
+
+* Add view log permission ([2b5ed50](https://git.fiotech.co/wbaindonesia/project-reporting/commit/2b5ed50e8ef6744ee5979a99565ee34b86e9c461))
+
+
+### Bug Fixes
+
+* Insert causer log as seeder when run db:seed ([7a257ee](https://git.fiotech.co/wbaindonesia/project-reporting/commit/7a257eeea7218697bc5aadab2294482805500448))
+
 ## 1.1.0-alpha.0 (2020-10-13)